first pass

+++ shared/src/db_mysql/util.rs
@@ -1,9 +1,9 @@
 use std::time::SystemTime;
 
 /// Creates unix timestamp.
-pub fn unix_timestamp() -> u64 {
+pub fn unix_timestamp() -> i64 {
   SystemTime::now()
     .duration_since(SystemTime::UNIX_EPOCH)
     .unwrap()
-    .as_secs()
+    .as_secs() as i64
 }
